Parties are assigned according to a defined threat level

+++ think-machine-random/src/main/java/com/softwaremagico/tm/random/party/RandomPartyMember.java
@@ -1,5 +1,29 @@
 package com.softwaremagico.tm.random.party;
 
+/*-
+ * #%L
+ * Think Machine (Random Generator)
+ * %%
+ * Copyright (C) 2017 - 2019 Softwaremagico
+ * %%
+ * This software is designed by Jorge Hortelano Otero. Jorge Hortelano Otero
+ * <softwaremagico@gmail.com> Valencia (Spain).
+ *  
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation; either version 2 of the License, or (at your option) any later
+ * version.
+ *  
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ * details.
+ *  
+ * You should have received a copy of the GNU General Public License along with
+ * this program; If not, see <http://www.gnu.org/licenses/gpl-3.0.html>.
+ * #L%
+ */
+
 import java.util.Set;
 
 import com.softwaremagico.tm.Element;
@@ -10,11 +34,11 @@ public class RandomPartyMember extends Element<RandomPartyMember> {
 	private final RandomProfile randomProfile;
 	private final Integer minNumber;
 	private final Integer maxNumber;
-	private final Float weight;
+	private final Integer weight;
 	private final Set<IRandomPreference> randomPreferences;
 
 	public RandomPartyMember(String id, String name, String language, RandomProfile randomProfile, Integer minNumber,
-			Integer maxNumber, Float weight, Set<IRandomPreference> randomPreferences) {
+			Integer maxNumber, Integer weight, Set<IRandomPreference> randomPreferences) {
 		super(id, name, language);
 		this.randomProfile = randomProfile;
 		this.minNumber = minNumber;
@@ -35,7 +59,7 @@ public class RandomPartyMember extends Element<RandomPartyMember> {
 		return maxNumber;
 	}
 
-	public Float getWeight() {
+	public Integer getWeight() {
 		return weight;
 	}
